api/requirements.txt
flask~=2.1.1
python-avatars~=1.3.0
mysql-connector-python==8.0.32